Rename: 게시물 저장 관련 파일 패키지 분리
