+++ src/main/java/com/greeny/ecomate/boardSave/dto/CreateBoardSaveRequestDto.java
@@ -1,11 +1,11 @@
-package com.greeny.ecomate.board.dto;
+package com.greeny.ecomate.boardSave.dto;
 
 import lombok.Data;
 
 import javax.validation.constraints.NotNull;
 
 @Data
-public class CreateSaveLogRequestDto {
+public class CreateBoardSaveRequestDto {
 
     @NotNull
     private Long boardId;
